events:  Refactor requirements to locals
This is a first step towards making events a bit more data driven, by
promoting the population and time requirements from being inlined in the
subclass functions, to being a local variable.

Eventually we'll move this stuff into an EventData.lua file, to follow
suit with how subclass data is handled everywhere else in the codebase.